sorting done

+++ pushes.c
@@ -6,7 +6,7 @@
 /*   By: adantas- <adantas-@student.42sp.org.br>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/02/09 12:48:51 by adantas-          #+#    #+#             */
-/*   Updated: 2023/02/14 15:28:12 by adantas-         ###   ########.fr       */
+/*   Updated: 2023/02/15 17:08:33 by adantas-         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,16 +14,16 @@
 
 void	push_a(t_stack *stks)
 {
-	size_t	i;
+	int	i;
 
 	if (!stks->sz_b)
 		return ;
-	i = stks->sz_a;
+	i = (int)stks->sz_a;
 	while (--i >= 0)
 		stks->a[i + 1] = stks->a[i];
 	stks->a[0] = stks->b[0];
 	i = -1;
-	while (++i < stks->sz_b - 1)
+	while (++i < (int)stks->sz_b - 1)
 		stks->b[i] = stks->b[i + 1];
 	stks->sz_a++;
 	stks->sz_b--;
@@ -32,16 +32,16 @@ void	push_a(t_stack *stks)
 
 void	push_b(t_stack *stks)
 {
-	size_t	i;
+	int	i;
 
 	if (!stks->sz_a)
 		return ;
-	i = stks->sz_b;
+	i = (int)stks->sz_b;
 	while (--i >= 0)
 		stks->b[i + 1] = stks->b[i];
 	stks->b[0] = stks->a[0];
 	i = -1;
-	while (++i < stks->sz_a - 1)
+	while (++i < (int)stks->sz_a - 1)
 		stks->a[i] = stks->a[i + 1];
 	stks->sz_a--;
 	stks->sz_b++;
